+++ src/main/java/org/folio/services/holding/HoldingsService.java
@@ -1,6 +1,7 @@
 package org.folio.services.holding;
 
 import static io.vertx.core.Promise.promise;
+import static org.apache.logging.log4j.LogManager.getLogger;
 import static org.folio.rest.impl.HoldingsStorageApi.HOLDINGS_RECORD_TABLE;
 import static org.folio.rest.impl.StorageHelper.MAX_ENTITIES;
 import static org.folio.rest.jaxrs.resource.HoldingsStorage.DeleteHoldingsStorageHoldingsByHoldingsRecordIdResponse;
@@ -14,28 +15,38 @@ import static org.folio.rest.persist.PgUtil.postSync;
 import static org.folio.rest.persist.PgUtil.postgresClient;
 import static org.folio.services.batch.BatchOperationContextFactory.buildBatchOperationContext;
 
+import io.vertx.core.CompositeFuture;
 import io.vertx.core.Context;
 import io.vertx.core.Future;
 import io.vertx.core.Promise;
 import io.vertx.core.json.JsonObject;
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import javax.ws.rs.core.Response;
 import org.apache.commons.lang3.StringUtils;
+import org.apache.logging.log4j.Logger;
 import org.folio.persist.HoldingsRepository;
+import org.folio.persist.InstanceInternalRepository;
 import org.folio.rest.jaxrs.model.HoldingsRecord;
 import org.folio.rest.persist.PostgresClient;
 import org.folio.rest.support.CqlQuery;
 import org.folio.rest.support.EndpointFailureHandler;
 import org.folio.rest.support.HridManager;
 import org.folio.rest.tools.utils.OptimisticLockingUtil;
+import org.folio.services.caches.ConsortiumData;
+import org.folio.services.caches.ConsortiumDataCache;
+import org.folio.services.consortium.ConsortiumService;
+import org.folio.services.consortium.ConsortiumServiceImpl;
+import org.folio.services.consortium.entities.SharingInstance;
 import org.folio.services.domainevent.HoldingDomainEventPublisher;
 import org.folio.services.domainevent.ItemDomainEventPublisher;
 import org.folio.validator.CommonValidators;
 import org.folio.validator.NotesValidators;
 
 public class HoldingsService {
+  private static final Logger log = getLogger(HoldingsService.class);
   private static final String INSTANCE_ID = "instanceId";
   private final Context vertxContext;
   private final Map<String, String> okapiHeaders;
@@ -44,6 +55,8 @@ public class HoldingsService {
   private final HoldingsRepository holdingsRepository;
   private final ItemDomainEventPublisher itemEventService;
   private final HoldingDomainEventPublisher domainEventPublisher;
+  private final InstanceInternalRepository instanceRepository;
+  private final ConsortiumService consortiumService;
 
   public HoldingsService(Context context, Map<String, String> okapiHeaders) {
     this.vertxContext = context;
@@ -54,6 +67,9 @@ public class HoldingsService {
     holdingsRepository = new HoldingsRepository(context, okapiHeaders);
     itemEventService = new ItemDomainEventPublisher(context, okapiHeaders);
     domainEventPublisher = new HoldingDomainEventPublisher(context, okapiHeaders);
+    instanceRepository = new InstanceInternalRepository(context, okapiHeaders);
+    consortiumService = new ConsortiumServiceImpl(context.owner().createHttpClient(),
+      context.get(ConsortiumDataCache.class.getName()));
   }
 
   /**
@@ -127,30 +143,34 @@ public class HoldingsService {
   }
 
   public Future<Response> createHoldings(List<HoldingsRecord> holdings, boolean upsert, boolean optimisticLocking) {
-    if (upsert) {
-      return upsertHoldings(holdings, optimisticLocking);
-    }
-
     for (HoldingsRecord holdingsRecord : holdings) {
       holdingsRecord.setEffectiveLocationId(calculateEffectiveLocation(holdingsRecord));
     }
 
-    return hridManager.populateHridForHoldings(holdings)
-      .compose(NotesValidators::refuseHoldingLongNotes)
-      .compose(result -> buildBatchOperationContext(upsert, holdings,
-        holdingsRepository, HoldingsRecord::getId))
-      .compose(batchOperation -> postSync(HOLDINGS_RECORD_TABLE, holdings, MAX_ENTITIES, upsert, optimisticLocking,
-        okapiHeaders, vertxContext, PostHoldingsStorageBatchSynchronousResponse.class)
-        .onSuccess(domainEventPublisher.publishCreatedOrUpdated(batchOperation)));
+    return consortiumService.getConsortiumData(okapiHeaders)
+      .compose(consortiumDataOptional -> {
+        if (consortiumDataOptional.isPresent()) {
+          return createShadowInstancesIfNeeded(holdings, consortiumDataOptional.get());
+        }
+        return Future.succeededFuture();
+      })
+      .compose(ar -> {
+        if (upsert) {
+          return upsertHoldings(holdings, optimisticLocking);
+        } else {
+          return hridManager.populateHridForHoldings(holdings)
+            .compose(NotesValidators::refuseHoldingLongNotes)
+            .compose(result -> buildBatchOperationContext(upsert, holdings,
+              holdingsRepository, HoldingsRecord::getId))
+            .compose(batchOperation -> postSync(HOLDINGS_RECORD_TABLE, holdings, MAX_ENTITIES,
+              upsert, optimisticLocking, okapiHeaders, vertxContext, PostHoldingsStorageBatchSynchronousResponse.class)
+              .onSuccess(domainEventPublisher.publishCreatedOrUpdated(batchOperation)));
+        }
+      });
   }
-
 
   public Future<Response> upsertHoldings(List<HoldingsRecord> holdings, boolean optimisticLocking) {
     try {
-      for (HoldingsRecord holdingsRecord : holdings) {
-        holdingsRecord.setEffectiveLocationId(calculateEffectiveLocation(holdingsRecord));
-      }
-
       if (optimisticLocking) {
         OptimisticLockingUtil.unsetVersionIfMinusOne(holdings);
       } else {
@@ -214,6 +234,29 @@ public class HoldingsService {
     } else {
       return permanentLocationId;
     }
+  }
+
+  private CompositeFuture createShadowInstancesIfNeeded(List<HoldingsRecord> holdingsRecords,
+                                                        ConsortiumData consortiumData) {
+    Map<String, Future<SharingInstance>> instanceFuturesMap = new HashMap<>();
+
+    for (HoldingsRecord holdingRecord : holdingsRecords) {
+      String instanceId = holdingRecord.getInstanceId();
+      instanceFuturesMap.computeIfAbsent(instanceId, v -> createShadowInstanceIfNeeded(instanceId, consortiumData));
+    }
+    return CompositeFuture.all(new ArrayList<>(instanceFuturesMap.values()));
+  }
+
+  private Future<SharingInstance> createShadowInstanceIfNeeded(String instanceId, ConsortiumData consortiumData) {
+    return instanceRepository.getById(instanceId)
+      .compose(instance -> {
+        if (instance != null) {
+          return Future.succeededFuture();
+        }
+        log.info("createShadowInstanceIfNeeded:: instance with id: {} is not found in local tenant."
+          + " Trying to create a shadow instance", instanceId);
+        return consortiumService.createShadowInstance(instanceId, consortiumData, okapiHeaders);
+      });
   }
 
   private boolean holdingsRecordFound(HoldingsRecord holdingsRecord) {