+++ src/test/java/org/folio/rest/api/TestBase.java
@@ -68,6 +68,7 @@ public abstract class TestBase {
   static ResourceClient illPoliciesClient;
   static ResourceClient inventoryViewClient;
   static ResourceClient statisticalCodeClient;
+  static ResourceClient boundWithClient;
   static StatisticalCodeFixture statisticalCodeFixture;
   static InstanceReindexFixture instanceReindex;
   static AsyncMigrationFixture asyncMigration;
@@ -91,6 +92,7 @@ public abstract class TestBase {
     itemsStorageSyncClient = ResourceClient.forItemsStorageSync(getClient());
     inventoryViewClient = ResourceClient.forInventoryView(getClient());
     statisticalCodeClient = ResourceClient.forStatisticalCodes(getClient());
+    boundWithClient = ResourceClient.forBoundWithParts(getClient());
     instancesStorageBatchInstancesClient = ResourceClient
       .forInstancesStorageBatchInstances(getClient());
     instanceTypesClient = ResourceClient.forInstanceTypes(getClient());
@@ -129,17 +131,21 @@ public abstract class TestBase {
    * accordingly.
    */
   protected static void clearData() {
-    StorageTestSuite.deleteAll(itemsStorageUrl(""));
-    StorageTestSuite.deleteAll(holdingsStorageUrl(""));
-    StorageTestSuite.deleteAll(instancesStorageUrl(""));
-    StorageTestSuite.deleteAll(locationsStorageUrl(""));
-    StorageTestSuite.deleteAll(locLibraryStorageUrl(""));
-    StorageTestSuite.deleteAll(locCampusStorageUrl(""));
-    StorageTestSuite.deleteAll(locInstitutionStorageUrl(""));
-    StorageTestSuite.deleteAll(loanTypesStorageUrl(""));
-    StorageTestSuite.deleteAll(materialTypesStorageUrl(""));
-    StorageTestSuite.deleteAll(servicePointsUsersUrl(""));
-    StorageTestSuite.deleteAll(servicePointsUrl(""));
+    clearData(TENANT_ID);
+  }
+
+  protected static void clearData(String tenantId) {
+    StorageTestSuite.deleteAll(itemsStorageUrl(""), tenantId);
+    StorageTestSuite.deleteAll(holdingsStorageUrl(""), tenantId);
+    StorageTestSuite.deleteAll(instancesStorageUrl(""), tenantId);
+    StorageTestSuite.deleteAll(locationsStorageUrl(""), tenantId);
+    StorageTestSuite.deleteAll(locLibraryStorageUrl(""), tenantId);
+    StorageTestSuite.deleteAll(locCampusStorageUrl(""), tenantId);
+    StorageTestSuite.deleteAll(locInstitutionStorageUrl(""), tenantId);
+    StorageTestSuite.deleteAll(loanTypesStorageUrl(""), tenantId);
+    StorageTestSuite.deleteAll(materialTypesStorageUrl(""), tenantId);
+    StorageTestSuite.deleteAll(servicePointsUsersUrl(""), tenantId);
+    StorageTestSuite.deleteAll(servicePointsUrl(""), tenantId);
   }
 
   /**
@@ -187,9 +193,13 @@ public abstract class TestBase {
    * @param url endpoint where to execute a GET request
    */
   void assertGetNotFound(URL url) {
+    assertGetNotFound(url, TENANT_ID);
+  }
+
+  void assertGetNotFound(URL url, String tenantId) {
     CompletableFuture<Response> getCompleted = new CompletableFuture<>();
 
-    getClient().get(url, TENANT_ID, ResponseHandler.text(getCompleted));
+    getClient().get(url, tenantId, ResponseHandler.text(getCompleted));
     Response response = get(getCompleted);
     assertThat(response.getStatusCode(), is(HttpURLConnection.HTTP_NOT_FOUND));
   }